Add input buffering and support for polling
The driver now supports a listening mode, which is set by writing to
the sysfs attribute 'listen'. In this mode, the radio is constantly
in RX mode and stores incoming packets in a receive buffer.

When not in listening mode, the behavior is the same as before.

poll() is also implemented to be able to use select() to check for
new packets.